% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC.R
\name{Carlson_RC}
\alias{Carlson_RC}
\title{Carlson elliptic integral RC}
\usage{
Carlson_RC(x, y, minerror = .Machine$double.eps)
}
\arguments{
\item{x, y}{real or complex numbers; at least...}

\item{minerror}{bound of relative error}
}
\value{
A complex number.
}
\description{
Evaluate the Carlson elliptic integral RC.
}
\examples{
Carlson_RC(5, 2)
gsl::ellint_RC(5, 2)
}
