% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF.R
\name{Carlson_RF}
\alias{Carlson_RF}
\title{Carlson elliptic integral RF}
\usage{
Carlson_RF(x, y, z, minerror = 2 * .Machine$double.eps)
}
\arguments{
\item{x, y, z}{real or complex numbers; at most one can be 0}

\item{minerror}{minimal relative error}
}
\value{
A complex number.
}
\description{
Evaluate the Carlson elliptic integral RF.
}
\examples{
Carlson_RF(5, 2, 3)
gsl::ellint_RF(5, 2, 3)
}
