% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RG.R
\name{Carlson_RG}
\alias{Carlson_RG}
\title{Carlson elliptic integral RG}
\usage{
Carlson_RG(x, y, z, minerror = 2 * .Machine$double.eps)
}
\arguments{
\item{x, y, z}{real or complex numbers; at most one can be zero}

\item{minerror}{bound on the relative error passed to
\code{\link{Carlson_RF}} and \code{\link{Carlson_RD}}}
}
\value{
A complex number.
}
\description{
Evaluate the Carlson elliptic integral RG.
}
