% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RJ.R
\name{Carlson_RJ}
\alias{Carlson_RJ}
\title{Carlson elliptic integral RJ}
\usage{
Carlson_RJ(x, y, z, p, minerror = 2 * .Machine$double.eps)
}
\arguments{
\item{x, y, z, p}{real or complex numbers; at least...}

\item{minerror}{minimal relative error}
}
\value{
A complex number.
}
\description{
Evaluate the Carlson elliptic integral RJ.
}
\examples{
Carlson_RJ(5, 2, 3, 4)
gsl::ellint_RJ(5, 2, 3, 4)
}
