% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptic_Z.R
\name{elliptic_Z}
\alias{elliptic_Z}
\title{Jacobi zeta function}
\usage{
elliptic_Z(phi, m, minerror = 2 * .Machine$double.eps)
}
\arguments{
\item{phi}{amplitude, real or complex number}

\item{m}{parameter, real or complex number}

\item{minerror}{bound on relative error passed to \code{\link{elliptic_E}}
and \code{\link{elliptic_F}}}
}
\value{
A complex number.
}
\description{
Evaluate the Jacobi zeta function.
}
