% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeumanLambda.R
\name{Lambda0}
\alias{Lambda0}
\title{Heuman Lambda function}
\usage{
Lambda0(phi, m, minerror = 1e-14)
}
\arguments{
\item{phi}{Jacobi amplitude, a complex number}

\item{m}{parameter, a complex number}

\item{minerror}{the bound on the relative error passed to
\code{\link{elliptic_F}} and \code{\link{elliptic_Z}}}
}
\value{
A complex number.
}
\description{
Evaluates the Heuman Lambda function.
}
