% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RD.R
\name{Carlson_RD}
\alias{Carlson_RD}
\title{Carlson elliptic integral RD}
\usage{
Carlson_RD(x, y, z, minerror = 2 * .Machine$double.eps)
}
\arguments{
\item{x, y, z}{real or complex numbers; at least...}

\item{minerror}{bound of relative error}
}
\value{
A complex number.
}
\description{
Evaluate the Carlson elliptic integral RD.
}
\examples{
Carlson_RD(5, 2, 3)
gsl::ellint_RD(5, 2, 3)
}
